Make `quote` and `unquote` operators special forms
Previously, the `quote` operator would return its argument wrapped in a QuotedExpressionNode, which prevented evaluation of the argument. Now, the `quote` operator is a special form that returns its argument unevaluated without creating a QuotedExpressionNode object. The QuotedExpressionNode class has been removed and the operators that took QuotedExpressionNode objects as arguments have been simplified. The `unquote` operator forces evaluation of its argument. It is not an identity function.